Rename recommended config to base
