+++ packages/eslint-plugin/lib/configs/base.js
